+++ klause/src/commonMain/kotlin/com/eignex/klause/lp/relaxation/SchedulingView.kt
@@ -20,7 +20,7 @@ import com.eignex.klause.solver.Problem
 internal class SchedulingView(
     val starts: IntArray,
     /** Constant per-task durations (`durations[i] == 0` tasks consume no resource). */
-    val durations: IntArray,
+    val durations: LongArray,
     /** Per-task minimum resource demand. */
     val resources: LongArray,
     /** Declared (maximum) capacity, `> 0`. */
@@ -35,13 +35,13 @@ internal fun schedulingViews(problem: Problem): List<SchedulingView> {
         when (f) {
             is Cumulative -> {
                 if (f.durationVars.isNotEmpty() || f.presents.isNotEmpty() || f.starts.isEmpty()) continue
-                val cap = if (f.capacityVar >= 0) problem.intDomains[f.capacityVar].max else f.capacity.toLong()
+                val cap = if (f.capacityVar >= 0) problem.intDomains[f.capacityVar].max else f.capacity
                 if (cap <= 0L) continue
                 val res = LongArray(f.starts.size) { i ->
                     if (f.resourceVars.isNotEmpty()) {
                         problem.intDomains[f.resourceVars[i]].min
                     } else {
-                        f.resources[i].toLong()
+                        f.resources[i]
                     }
                 }
                 out.add(SchedulingView(f.starts, f.durations, res, cap))
